Introduce high-level MemcachedClient
Create a high-level client that interacts with multiple Memcached servers
depending on the key or operation being performed. The client uses rendezvous
hashing to determine which server is responsible for any relevant keys.

In addition the following improvements have been made:
* `Key` type introduced to validate keys before using them
* `mc` has been rewritten to clean up resources properly

The following things are missing and will done in follow up work:
* Tests for newly introduced client and hashing
* Type-safe alternative to string hostnames