Use SocketAddr within Server abstraction
Rework the way `Server` in the client works such that it wraps an address
and server name instead of a string used for both connection and TLS name
resolution. This change also indexes server-specific results (stats, meta,
etc.) by ServerID instead of Server. ServerID is an opaque unique ID for
a particular server.

Fixes #103