Implementation of jump consistent hashing for selecting servers
This change adds new logic for selecting which server owns a particular
key based on the jump consistent hash. This logic is much faster than
the existing rendezvous hashing but is more operationally fragile.

Using jump consistent hashing requires that servers are always in the
same order and that servers are added and removed from the end of this
list, think how statefulsets in Kubernetes work.

It's not currently possible to make use of this in mtop since we have
no way to look up the names of servers until we add support for SRV
DNS records. Until then, servers are ordered by their SocketAddr which
means they would end up being removed from the middle of the list
which completely breaks jump consistent hashing.

Depends on #107