Fix tab overflow from lots of hosts
Allow all host tabs to be displayed even when there are more hosts than
can fit on the screen. This fix is a bit of a cheat since we aren't
actually moving the tabs, we just change the order of hosts based on
user input. If this proves annoying to use we can look at actually
moving the tabs around.

Fixes #117