Implement a rudimentary DNS client for SRV record lookup
This change creates a very basic DNS client capable of performing A, AAAA,
CNAME, NS, SOA, TXT, and SRV types of queries. This is the first step in
supporting DNS lookup of SRV records for service discovery in `mtop` and
`mc`. It also adds a `dns` executable capable of making queries and parsing
binary DNS messages.

This client has undergone very little testing and should be considered
both experimental and internal to mtop.

See #107