Read DNS settings from a resolv.conf file
Allow some DNS settings to be read from a system's /etc/resolv.conf
file. Noteably, only the `nameserver` and a few `option`s are supported
for now. This change also switches resolution of A and AAAA names to
our DNS client instead of using the system resolver via `lookup_host`.

Part of #107