dns: initial opt record support
This change introduces the ability to parse OPT records but
does not yet add them to requests or uses them when they are
part of a response. A future PR will start making use of OPT
records.

This change also does some unrelated cleanup, removing `--dns-local`
flags from each of the binaries in favor of picking this value
automatically. Additionally, the `--timeout-secs` option is removed
from the `dns` binary since it has been unused since resolv.conf
support was added.

Part of #107