Don't attempt to resolve bare hostnames
When connecting to Memcached servers, only attempt to resolve the hostnames
to IP addresses when prefixed with `dns+`. If just a hostname is supplied,
attempt to use it verbatim. This results in a better user experience since
we get the benefits of using the system resolver and the expected hostname
appears in the UI (instead of its IP).

Fixes #151