Add unit tests for MemcachedClient
Refactor MemcachedClient to make it unit testable and then add tests
of a few representative methods. This change makes extracts the `Selector`
logic for picking a server based on key to a trait and makes that type
generic in the client. This is more verbose but allows us to use
custom selector logic in the unit tests.