ci: Update ubuntu runners used by cargo dist
Update the ubuntu runners used by cargo dist to non-EOL versions
and move fuzzing and benchmarks to their own workspaces. This avoids
pulling in their dependencies when building and testing the rest of
the packages in the project. Fuzzing is still run as part of CI, it
just doesn't get built from the workspace root.